Move Perturbation
No need to be in strata-basics
Also move ShiftType
