+++ examples/src/main/java/com/opengamma/strata/examples/finance/HistoricalScenarioExample.java
@@ -45,11 +45,11 @@ import com.opengamma.strata.function.marketdata.curve.CurvePointShifts;
 import com.opengamma.strata.function.marketdata.curve.CurvePointShiftsBuilder;
 import com.opengamma.strata.function.marketdata.scenario.curve.AnyDiscountCurveFilter;
 import com.opengamma.strata.function.marketdata.scenario.curve.CurveRateIndexFilter;
+import com.opengamma.strata.market.ShiftType;
 import com.opengamma.strata.market.curve.Curve;
 import com.opengamma.strata.market.curve.CurveGroupName;
 import com.opengamma.strata.market.curve.CurveParameterMetadata;
 import com.opengamma.strata.market.curve.NodalCurve;
-import com.opengamma.strata.market.curve.perturb.ShiftType;
 import com.opengamma.strata.market.id.DiscountCurveId;
 import com.opengamma.strata.market.id.RateCurveId;
 import com.opengamma.strata.market.id.RateIndexCurveId;
