+++ modules/function/src/main/java/com/opengamma/strata/function/marketdata/curve/CurvePointShifts.java
@@ -35,10 +35,10 @@ import com.opengamma.strata.collect.Messages;
 import com.opengamma.strata.collect.array.DoubleArray;
 import com.opengamma.strata.collect.array.DoubleMatrix;
 import com.opengamma.strata.collect.tuple.ObjIntPair;
+import com.opengamma.strata.market.ShiftType;
 import com.opengamma.strata.market.curve.Curve;
 import com.opengamma.strata.market.curve.CurveParameterMetadata;
 import com.opengamma.strata.market.curve.NodalCurve;
-import com.opengamma.strata.market.curve.perturb.ShiftType;
 
 /**
  * A perturbation that applies different shifts to specific points on a curve.
