+++ modules/market/src/main/java/com/opengamma/strata/market/curve/perturb/CurveParallelShift.java
@@ -24,7 +24,8 @@ import org.joda.beans.impl.direct.DirectMetaPropertyMap;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import com.opengamma.strata.basics.market.Perturbation;
+import com.opengamma.strata.market.Perturbation;
+import com.opengamma.strata.market.ShiftType;
 import com.opengamma.strata.market.curve.Curve;
 
 /**
