+++ modules/market/src/test/java/com/opengamma/strata/market/curve/perturb/ParallelShiftedCurveTest.java
@@ -14,6 +14,7 @@ import static org.assertj.core.api.Assertions.offset;
 import org.testng.annotations.Test;
 
 import com.opengamma.strata.collect.array.DoubleArray;
+import com.opengamma.strata.market.ShiftType;
 import com.opengamma.strata.market.curve.ConstantNodalCurve;
 import com.opengamma.strata.market.curve.Curve;
 import com.opengamma.strata.market.curve.CurveMetadata;
