+++ modules/market/src/test/java/com/opengamma/strata/market/value/SimpleDiscountFactorsTest.java
@@ -21,8 +21,8 @@ import java.time.LocalDate;
 import org.testng.annotations.Test;
 
 import com.opengamma.strata.basics.interpolator.CurveInterpolator;
-import com.opengamma.strata.basics.market.Perturbation;
 import com.opengamma.strata.collect.array.DoubleArray;
+import com.opengamma.strata.market.Perturbation;
 import com.opengamma.strata.market.curve.Curve;
 import com.opengamma.strata.market.curve.CurveMetadata;
 import com.opengamma.strata.market.curve.CurveName;
