+++ modules/pricer/src/main/java/com/opengamma/strata/pricer/impl/credit/isda/FiniteDifferenceSpreadSensitivityCalculator.java
@@ -8,7 +8,7 @@ package com.opengamma.strata.pricer.impl.credit.isda;
 import java.util.Arrays;
 
 import com.opengamma.strata.collect.ArgChecker;
-import com.opengamma.strata.market.curve.perturb.ShiftType;
+import com.opengamma.strata.market.ShiftType;
 import com.opengamma.strata.math.impl.differentiation.FiniteDifferenceType;
 
 /**
