+++ modules/pricer/src/test/java/com/opengamma/strata/pricer/impl/credit/isda/CS01FromParSpreadTest.java
@@ -11,7 +11,7 @@ import java.time.LocalDate;
 
 import org.testng.annotations.Test;
 
-import com.opengamma.strata.market.curve.perturb.ShiftType;
+import com.opengamma.strata.market.ShiftType;
 
 /**
  * Test.
