+++ modules/pricer/src/test/java/com/opengamma/strata/pricer/impl/credit/isda/CS01TestGrids.java
@@ -18,7 +18,7 @@ import java.util.Arrays;
 import org.testng.annotations.Test;
 
 import com.opengamma.strata.collect.ArgChecker;
-import com.opengamma.strata.market.curve.perturb.ShiftType;
+import com.opengamma.strata.market.ShiftType;
 
 /**
  * Test.
