+++ modules/pricer/src/test/java/com/opengamma/strata/pricer/impl/credit/isda/SpreadSensitivityTest.java
@@ -17,7 +17,7 @@ import org.testng.annotations.Test;
 import com.opengamma.strata.basics.date.HolidayCalendar;
 import com.opengamma.strata.basics.date.HolidayCalendars;
 import com.opengamma.strata.basics.schedule.StubConvention;
-import com.opengamma.strata.market.curve.perturb.ShiftType;
+import com.opengamma.strata.market.ShiftType;
 import com.opengamma.strata.math.impl.differentiation.FiniteDifferenceType;
 
 /**
